improvements to line wrapping on import
