+++ frontend/src/components/import/import.module.scss
@@ -41,26 +41,51 @@
 
 .candidateHeader {
     display: flex;
-    gap: 1rem;
+    column-gap: 1rem;
+    row-gap: 0.5rem;
     align-items: top;
     margin-left: 0.5rem;
+    margin-top: 0.1rem;
+    margin-bottom: 0.8rem;
     width: 100%;
+    flex-wrap: wrap;
 
     .headerGroup {
         display: flex;
         gap: 0.5rem;
         align-items: center;
+        flex-wrap: wrap;
 
         > [data-changed="true"] {
             color: #ebcb8c;
         }
     }
 
+    // not working yet! candidate header is overflowing its container
+    // and i did not manage to fix that.
+    .artstAlbumLabel {
+        overflow: hidden;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+    }
+
     .penaltyIconRow {
         display: flex;
         gap: 0.25rem;
         align-items: center;
+    }
+
+    .penaltyIconRowWrapper {
+        flex-basis: auto;
         margin-left: auto;
+
+        @media (max-width: 800px) {
+            display: none;
+            // gets too crowded, imho.
+            // width: 100%;
+            // margin-left: 0.9rem;
+            // margin-bottom: 0.2rem;
+        }
     }
 }
